simple_cf/requirements.txt
google-cloud-bigquery
pytest~=6.2.2
